refactor(cli): extract project and task commands
